Update coda_m.py to include folder monitor functionality.
I've implemented the first go at a command to process data as it is collected. Note that this has yet to be tested, so there are likely to be bugs I haven't foreseen (I suspect there could also be an issue where partial files get added to the monitored folder, but I'll have to look into that).